+++ Pandora/Paradigm/Structure/Ability/Substructure.hs
@@ -7,7 +7,7 @@ import Pandora.Pattern.Category ((.))
 import Pandora.Pattern.Functor.Covariant (comap)
 import Pandora.Pattern.Functor.Extractable (extract)
 import Pandora.Paradigm.Controlflow.Effect.Interpreted (run)
-import Pandora.Paradigm.Inventory.Optics (type (:~.), view, over)
+import Pandora.Paradigm.Inventory.Optics (type (:~.), view, over, set)
 import Pandora.Paradigm.Primary.Functor.Tagged (Tagged (Tag))
 import Pandora.Paradigm.Schemes.TU (TU (TU), type (<:.>))
 
@@ -24,6 +24,10 @@ subview = view (sub @f)
 substitute :: forall f t . Substructure f t
 	=> (Substructural f t ~> Substructural f t) -> t ~> t
 substitute = over (sub @f)
+
+subplace :: forall f t a . Substructure f t
+	=> (Substructural f t) a -> t a -> t a
+subplace = set (sub @f)
 
 data Command a = Delete a